+++ build.sh
@@ -1,6 +1,0 @@
-go get -u github.com/DanielMabbett/gojenkins
-go get -u github.com/bndr/gojenkins
-go get -u github.com/DanielMabbett/terraform-provider-jenkinsci/jenkinsci
-go get -u github.com/hashicorp/terraform/helper/schema
-
-go build -o terraform-provider-jenkinsci