+++ jenkinsci/resource_jenkinsci_folder.go
@@ -1,6 +1,7 @@
 package jenkinsci
 
 import (
+	"fmt"
 	jenkins "github.com/DanielMabbett/gojenkins"
 	"github.com/hashicorp/terraform/helper/schema"
 )
@@ -34,21 +35,21 @@ func resourceFolderCreate(d *schema.ResourceData, meta interface{}) error {
 	name := d.Get("name").(string)
 
 	if _, ok := d.GetOk("parent_folder"); ok {
-		// Create a nested folder
 		parentFolder := d.Get("parent_folder").(string)
-		
+
 		nFolder, err := client.CreateFolder(name, parentFolder)
 		if err != nil {
-			panic(err)
+			return fmt.Errorf("Error creating the Jenkins Folder: %s", err)
 		}
 		d.SetId(nFolder.GetName())
+
 	} else {
-		// Create folder normally
 		pFolder, err := client.CreateFolder(name)
 		if err != nil {
-			panic(err)
+			return fmt.Errorf("Error creating the Jenkins Folder: %s", err)
 		}
 		d.SetId(pFolder.GetName())
+
 	}
 
 	return resourceFolderRead(d, meta)