+++ jenkinsci/resource_jenkinsci_view.go
@@ -1,8 +1,8 @@
 package jenkinsci
 
 import (
+	"fmt"
 	jenkins "github.com/DanielMabbett/gojenkins"
-	// jenkins "github.com/danielmabbett/gojenkins"
 	"github.com/hashicorp/terraform/helper/schema"
 )
 
@@ -10,17 +10,23 @@ func resourceView() *schema.Resource {
 	return &schema.Resource{
 		Create: resourceViewCreate,
 		Read:   resourceViewRead,
-		Update: resourceViewUpdate,
+		// Update: resourceViewUpdate,
 		Delete: resourceViewDelete,
 
 		Schema: map[string]*schema.Schema{
 			"name": {
 				Type:     schema.TypeString,
 				Required: true,
+				ForceNew: true,
 			},
-			"assigned_project": {
-				Type:     schema.TypeString,
+
+			"assigned_projects": {
+				Type:     schema.TypeList,
 				Optional: true,
+				ForceNew: true,
+				Elem: &schema.Schema{
+					Type: schema.TypeString,
+				},
 			},
 		},
 	}
@@ -32,12 +38,12 @@ func resourceViewCreate(d *schema.ResourceData, meta interface{}) error {
 
 	view, err := client.CreateView(name, jenkins.LIST_VIEW)
 	if err != nil {
-		panic(err)
+		return fmt.Errorf("Error creating the Jenkins View: %s", err)
 	}
 
-	if _, ok := d.GetOk("assigned_project"); ok {
-		assignedProject := d.Get("assigned_project").(string)
-		view.AddJob(assignedProject)
+	assigedProjects := d.Get("assigned_projects").([]interface{})
+	for _, project := range assigedProjects {
+		view.AddJob(project.(string))
 	}
 
 	d.SetId(view.GetName())
@@ -50,27 +56,31 @@ func resourceViewRead(d *schema.ResourceData, meta interface{}) error {
 
 func resourceViewUpdate(d *schema.ResourceData, meta interface{}) error {
 
-	if d.HasChange("assigned_project") {
-		client := meta.(*jenkins.Jenkins)
-		oldProj, newProj := d.GetChange("assigned_project")
-
-		v, err := client.GetView(d.Get("assigned_project").(string))
-		if err != nil {
-			panic(err)
-		}
-
-		v.DeleteJob(oldProj.(string))
-		v.AddJob(newProj.(string))
-
-		d.SetPartial("assigned_project")
-	}
+	// 	if d.HasChange("assigned_project") {
+	// 		client := meta.(*jenkins.Jenkins)
+	// 		oldProj, newProj := d.GetChange("assigned_project")
+	//
+	// 		v, err := client.GetView(d.Get("assigned_project").(string))
+	// 		if err != nil {
+	// 			return fmt.Errorf("Error getting the Jenkins View: %s", err)
+	// 		}
+	//
+	// 		v.DeleteJob(oldProj.(string))
+	// 		v.AddJob(newProj.(string))
+	//
+	// 		d.SetPartial("assigned_project")
+	// 	}
 	return resourceViewRead(d, meta)
 }
 
 func resourceViewDelete(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(*jenkins.Jenkins)
 	name := d.Get("name").(string)
-	fullPath := "/view/" + name
-	client.DeleteJob(fullPath)
+
+	_, err := client.DeleteView(name)
+	if err != nil {
+		return fmt.Errorf("Error deleting the Jenkins View: %s", err)
+	}
+
 	return nil
 }