+++ test/provider.tf
@@ -1,5 +1,0 @@
-provider "jenkinsci" {
-  jenkins_endpoint       = "http://localhost:8080"
-  jenkins_admin_username = "jenkins"
-  jenkins_admin_password = ""
-}